Move enum to SearchBeatmapSetsRequest
